move commands to main
