+++ taotip/requirements.txt
@@ -8,3 +8,4 @@ cryptography>=3.1.1
 git+https://github.com/opentensor/bittensor.git#bittensor@master
 tqdm>=4.27,<4.50.0
 discord-py-interactions>=4.3.1
+py-cord>=2.1.1
